+++ src/main/java/com/atlauncher/gui/tabs/settings/ModsSettingsTab.java
@@ -51,7 +51,8 @@ public class ModsSettingsTab extends AbstractSettingsTab {
         gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
 
         JLabelWithHover defaultModPlatformLabel = new JLabelWithHover(GetText.tr("Default Mod Platform") + ":",
-                HELP_ICON, GetText.tr("The default mod platform to use when adding mods to instances."));
+                HELP_ICON, GetText.tr(
+                        "The default mod platform to use when adding mods to instances, as well as the platform to use when updating/reinstalling mods on multiple platforms."));
 
         add(defaultModPlatformLabel, gbc);